Issue#523: Refactor XML and JSON handling in CRUD operations
