+++ bundles/sirix-rest-api/src/main/kotlin/io/sirix/rest/crud/json/JsonCreate.kt
@@ -24,7 +24,6 @@ import io.sirix.service.json.shredder.JsonShredder
 import java.io.StringWriter
 import java.nio.file.Path
 
-private const val MAX_NODES_TO_SERIALIZE = 5000
 
 class JsonCreate(
     location: Path,
@@ -92,20 +91,6 @@ class JsonCreate(
         )
     }
 
-    override suspend fun createDatabaseIfNotExists(
-        dbFile: Path,
-        context: Context
-    ): DatabaseConfiguration? {
-        val dbConfig = prepareDatabasePath(dbFile, context)
-        return context.executeBlocking { promise: Promise<DatabaseConfiguration> ->
-            if (!Databases.existsDatabase(dbFile)) {
-                Databases.createJsonDatabase(dbConfig)
-            }
-
-            promise.complete(dbConfig)
-        }.await()
-    }
-
 
     private suspend fun insertJsonSubtreeAsFirstChild(
         manager: JsonResourceSession,
@@ -147,5 +132,9 @@ class JsonCreate(
 
     override suspend fun openDatabase(dbFile: Path, sirixDBUser: User): Database<JsonResourceSession> {
         return Databases.openJsonDatabase(dbFile, sirixDBUser)
+    }
+
+    override fun createDatabase(dbConfig: DatabaseConfiguration?) {
+        Databases.createJsonDatabase(dbConfig)
     }
 }
