+++ bundles/sirix-rest-api/src/main/kotlin/io/sirix/rest/crud/json/JsonDelete.kt
@@ -16,26 +16,13 @@ import io.sirix.query.json.BasicJsonDBStore
 import java.nio.file.Path
 
 class JsonDelete(location: Path, private val authz: AuthorizationProvider) : AbstractDeleteHandler(location) {
-    suspend fun handle(ctx: RoutingContext): Route {
-        val databaseName: String? = ctx.pathParam("database") ?: ctx.get("databaseName")
-        val resource: String? = ctx.pathParam("resource")
-        val nodeId: String? = ctx.queryParam("nodeId").getOrNull(0)
-
-        if (databaseName == null) {
-            dropDatabasesOfType(ctx, DatabaseType.JSON)
-        } else {
-            delete(databaseName, resource, nodeId?.toLongOrNull(), ctx)
-        }
-
-        return ctx.currentRoute()
-    }
 
     override fun createStore(ctx: RoutingContext): StructuredItemStore {
         return JsonSessionDBStore(ctx, BasicJsonDBStore.newBuilder().build(), ctx.get("user") as User, authz)
     }
 
     override fun database(dbFile: Path, sirixDBUser: io.sirix.access.User): Database<*> {
-        return Databases.openJsonDatabase(dbFile,  sirixDBUser)
+        return Databases.openJsonDatabase(dbFile, sirixDBUser)
     }
 
     override fun hashType(manager: ResourceSession<*, *>): HashType {
@@ -44,4 +31,6 @@ class JsonDelete(location: Path, private val authz: AuthorizationProvider) : Abs
         else
             throw IllegalArgumentException("Resource manager is not of JSON type.")
     }
+
+    override fun getDatabaseType(): DatabaseType = DatabaseType.JSON
 }
