+++ bundles/sirix-rest-api/src/main/kotlin/io/sirix/rest/crud/xml/XmlDelete.kt
@@ -2,7 +2,6 @@ package io.sirix.rest.crud.xml
 
 import io.vertx.ext.auth.User
 import io.vertx.ext.auth.authorization.AuthorizationProvider
-import io.vertx.ext.web.Route
 import io.vertx.ext.web.RoutingContext
 import io.brackit.query.jdm.StructuredItemStore
 import io.sirix.access.DatabaseType
@@ -16,26 +15,13 @@ import io.sirix.query.node.BasicXmlDBStore
 import java.nio.file.Path
 
 class XmlDelete(location: Path, private val authz: AuthorizationProvider) : AbstractDeleteHandler(location) {
-    suspend fun handle(ctx: RoutingContext): Route {
-        val databaseName: String? = ctx.pathParam("database") ?: ctx.get("databaseName")
-        val resource: String? = ctx.pathParam("resource")
-        val nodeId: String? = ctx.queryParam("nodeId").getOrNull(0)
-
-        if (databaseName == null) {
-            dropDatabasesOfType(ctx, DatabaseType.XML)
-        } else {
-            delete(databaseName, resource, nodeId?.toLongOrNull(), ctx)
-        }
-
-        return ctx.currentRoute()
-    }
 
     override fun createStore(ctx: RoutingContext): StructuredItemStore {
         return XmlSessionDBStore(ctx, BasicXmlDBStore.newBuilder().build(), ctx.get("user") as User, authz)
     }
 
     override fun database(dbFile: Path, sirixDBUser: io.sirix.access.User): Database<*> {
-        return Databases.openXmlDatabase(dbFile,  sirixDBUser)
+        return Databases.openXmlDatabase(dbFile, sirixDBUser)
     }
 
     override fun hashType(manager: ResourceSession<*, *>): HashType {
@@ -44,4 +30,6 @@ class XmlDelete(location: Path, private val authz: AuthorizationProvider) : Abst
         else
             throw IllegalArgumentException("Resource manager is not of XML type.")
     }
+
+    override fun getDatabaseType(): DatabaseType = DatabaseType.XML
 }
